# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: string;
  amount: number;
  type: TransactionType;
  categoryId: string;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  note?: string;
  date: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTransactionInputType =
  (typeof CreateTransactionInputType)[keyof typeof CreateTransactionInputType];

export const CreateTransactionInputType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateTransactionInput {
  amount: number;
  type: CreateTransactionInputType;
  categoryId: string;
  note?: string;
  date: string;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface Category {
  id: string;
  name: string;
  icon: string;
  color: string;
  type: CategoryType;
  createdAt: string;
}

export type CreateCategoryInputType =
  (typeof CreateCategoryInputType)[keyof typeof CreateCategoryInputType];

export const CreateCategoryInputType = {
  income: "income",
  expense: "expense",
  both: "both",
} as const;

export interface CreateCategoryInput {
  name: string;
  icon: string;
  color: string;
  type: CreateCategoryInputType;
}

export type BudgetPeriod = (typeof BudgetPeriod)[keyof typeof BudgetPeriod];

export const BudgetPeriod = {
  monthly: "monthly",
  weekly: "weekly",
  yearly: "yearly",
} as const;

export interface Budget {
  id: string;
  categoryId: string;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  amount: number;
  spent: number;
  period: BudgetPeriod;
  createdAt: string;
  updatedAt: string;
}

export type CreateBudgetInputPeriod =
  (typeof CreateBudgetInputPeriod)[keyof typeof CreateBudgetInputPeriod];

export const CreateBudgetInputPeriod = {
  monthly: "monthly",
  weekly: "weekly",
  yearly: "yearly",
} as const;

export interface CreateBudgetInput {
  categoryId: string;
  amount: number;
  period: CreateBudgetInputPeriod;
}

export type FinancialSummaryExpenseByCategoryItem = {
  categoryId: string;
  categoryName: string;
  categoryColor: string;
  categoryIcon: string;
  amount: number;
  percentage: number;
};

export interface FinancialSummary {
  totalIncome: number;
  totalExpense: number;
  balance: number;
  expenseByCategory: FinancialSummaryExpenseByCategoryItem[];
  recentTransactions: Transaction[];
}

export type GetTransactionsParams = {
  categoryId?: string;
  type?: GetTransactionsType;
  startDate?: string;
  endDate?: string;
};

export type GetTransactionsType =
  (typeof GetTransactionsType)[keyof typeof GetTransactionsType];

export const GetTransactionsType = {
  income: "income",
  expense: "expense",
} as const;

export type GetSummaryParams = {
  month?: string;
  year?: string;
};