# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: transactions
    description: Transaction operations
  - name: categories
    description: Category operations
  - name: budgets
    description: Budget operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get all transactions
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income, expense]
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionInput"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /budgets:
    get:
      operationId: getBudgets
      tags: [budgets]
      summary: Get all budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Budget"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetInput"
      responses:
        "201":
          description: Created budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"

  /budgets/{id}:
    put:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetInput"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /summary:
    get:
      operationId: getSummary
      tags: [transactions]
      summary: Get financial summary
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
        - name: year
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Financial summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Transaction:
      type: object
      properties:
        id:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        categoryId:
          type: string
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        note:
          type: string
        date:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - amount
        - type
        - categoryId
        - date
        - createdAt
        - updatedAt

    CreateTransactionInput:
      type: object
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        categoryId:
          type: string
        note:
          type: string
        date:
          type: string
      required:
        - amount
        - type
        - categoryId
        - date

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
        createdAt:
          type: string
      required:
        - id
        - name
        - icon
        - color
        - type
        - createdAt

    CreateCategoryInput:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        color:
          type: string
        type:
          type: string
          enum: [income, expense, both]
      required:
        - name
        - icon
        - color
        - type

    Budget:
      type: object
      properties:
        id:
          type: string
        categoryId:
          type: string
        categoryName:
          type: string
        categoryColor:
          type: string
        categoryIcon:
          type: string
        amount:
          type: number
        spent:
          type: number
        period:
          type: string
          enum: [monthly, weekly, yearly]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - categoryId
        - amount
        - spent
        - period
        - createdAt
        - updatedAt

    CreateBudgetInput:
      type: object
      properties:
        categoryId:
          type: string
        amount:
          type: number
        period:
          type: string
          enum: [monthly, weekly, yearly]
      required:
        - categoryId
        - amount
        - period

    FinancialSummary:
      type: object
      properties:
        totalIncome:
          type: number
        totalExpense:
          type: number
        balance:
          type: number
        expenseByCategory:
          type: array
          items:
            type: object
            properties:
              categoryId:
                type: string
              categoryName:
                type: string
              categoryColor:
                type: string
              categoryIcon:
                type: string
              amount:
                type: number
              percentage:
                type: number
            required:
              - categoryId
              - categoryName
              - categoryColor
              - categoryIcon
              - amount
              - percentage
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
      required:
        - totalIncome
        - totalExpense
        - balance
        - expenseByCategory
        - recentTransactions