# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetPeriod } from "./budgetPeriod";

export interface Budget {
  id: string;
  categoryId: string;
  categoryName?: string;
  categoryColor?: string;
  categoryIcon?: string;
  amount: number;
  spent: number;
  period: BudgetPeriod;
  createdAt: string;
  updatedAt: string;
}