# lib/api-zod/src/generated/types/createBudgetInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Personal Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBudgetInputPeriod } from "./createBudgetInputPeriod";

export interface CreateBudgetInput {
  categoryId: string;
  amount: number;
  period: CreateBudgetInputPeriod;
}